implementation; part 2
