+++ ginac/normal.cpp
@@ -125,6 +125,199 @@ numeric mul::integer_content() const
 }
 
 
+/** Rationalization of non-rational functions.
+ *  This function converts a general expression to a rational function
+ *  by replacing all non-rational subexpressions (like non-rational numbers,
+ *  non-integer powers or functions like sin(), cos() etc.) to temporary
+ *  symbols. This makes it possible to use functions like gcd() and divide()
+ *  on non-rational functions by applying to_rational() on the arguments,
+ *  calling the desired function and re-substituting the temporary symbols
+ *  in the result. To make the last step possible, all temporary symbols and
+ *  their associated expressions are collected in the map specified by the
+ *  repl parameter, ready to be passed as an argument to ex::subs().
+ *
+ *  @param repl collects all temporary symbols and their replacements
+ *  @return rationalized expression */
+ex ex::to_rational(exmap & repl) const
+{
+	return bp->to_rational(repl);
+}
+
+// GiNaC 1.1 compatibility function
+ex ex::to_rational(lst & repl_lst) const
+{
+	// Convert lst to exmap
+	exmap m;
+	for (const auto & elem : repl_lst)
+		m.insert(std::make_pair(elem.op(0), elem.op(1)));
+
+	ex ret = bp->to_rational(m);
+
+	// Convert exmap back to lst
+	repl_lst.remove_all();
+	for (const auto& elem : m)
+		repl_lst.append(elem.first == elem.second);
+
+	return ret;
+}
+
+ex ex::to_polynomial(exmap & repl) const
+{
+	return bp->to_polynomial(repl);
+}
+
+// GiNaC 1.1 compatibility function
+ex ex::to_polynomial(lst & repl_lst) const
+{
+	// Convert lst to exmap
+	exmap m;
+	for (const auto & elem : repl_lst)
+		m.insert(std::make_pair(elem.op(0), elem.op(1)));
+
+	ex ret = bp->to_polynomial(m);
+
+	// Convert exmap back to lst
+	repl_lst.remove_all();
+	for (const auto& elem : m)
+		repl_lst.append(elem.first == elem.second);
+
+	return ret;
+}
+
+/** Default implementation of ex::to_rational(). This replaces the object with
+ *  a temporary symbol. */
+ex basic::to_rational(exmap & repl) const
+{
+	return replace_with_symbol(*this, repl);
+}
+
+ex basic::to_polynomial(exmap & repl) const
+{
+	return replace_with_symbol(*this, repl);
+}
+
+
+/** Implementation of ex::to_rational() for symbols. This returns the
+ *  unmodified symbol. */
+ex symbol::to_rational(exmap & repl) const
+{
+	return *this;
+}
+
+/** Implementation of ex::to_polynomial() for symbols. This returns the
+ *  unmodified symbol. */
+ex symbol::to_polynomial(exmap & repl) const
+{
+	return *this;
+}
+
+
+/** Implementation of ex::to_rational() for a numeric. It splits complex
+ *  numbers into re+I*im and replaces I and non-rational real numbers with a
+ *  temporary symbol. */
+ex numeric::to_rational(exmap & repl) const
+{
+	if (is_real()) {
+		if (!is_rational())
+			return replace_with_symbol(*this, repl);
+	} else { // complex
+		numeric re = real();
+		numeric im = imag();
+		ex re_ex = re.is_rational() ? re : replace_with_symbol(re, repl);
+		ex im_ex = im.is_rational() ? im : replace_with_symbol(im, repl);
+		return re_ex + im_ex * replace_with_symbol(I, repl);
+	}
+	return *this;
+}
+
+/** Implementation of ex::to_polynomial() for a numeric. It splits complex
+ *  numbers into re+I*im and replaces I and non-integer real numbers with a
+ *  temporary symbol. */
+ex numeric::to_polynomial(exmap & repl) const
+{
+	if (is_real()) {
+		if (!is_integer())
+			return replace_with_symbol(*this, repl);
+	} else { // complex
+		numeric re = real();
+		numeric im = imag();
+		ex re_ex = re.is_integer() ? re : replace_with_symbol(re, repl);
+		ex im_ex = im.is_integer() ? im : replace_with_symbol(im, repl);
+		return re_ex + im_ex * replace_with_symbol(I, repl);
+	}
+	return *this;
+}
+
+
+/** Implementation of ex::to_rational() for powers. It replaces non-integer
+ *  powers by temporary symbols. */
+ex power::to_rational(exmap & repl) const
+{
+	if (exponent.info(info_flags::integer))
+		return power(basis.to_rational(repl), exponent);
+	else
+		return replace_with_symbol(*this, repl);
+}
+
+/** Implementation of ex::to_polynomial() for powers. It replaces non-posint
+ *  powers by temporary symbols. */
+ex power::to_polynomial(exmap & repl) const
+{
+	if (exponent.info(info_flags::posint))
+		return power(basis.to_rational(repl), exponent);
+	else if (exponent.info(info_flags::negint))
+	{
+		ex basis_pref = collect_common_factors(basis);
+		if (is_exactly_a<mul>(basis_pref) || is_exactly_a<power>(basis_pref)) {
+			// (A*B)^n will be automagically transformed to A^n*B^n
+			ex t = power(basis_pref, exponent);
+			return t.to_polynomial(repl);
+		}
+		else
+			return power(replace_with_symbol(power(basis, _ex_1), repl), -exponent);
+	} 
+	else
+		return replace_with_symbol(*this, repl);
+}
+
+
+/** Implementation of ex::to_rational() for expairseqs. */
+ex expairseq::to_rational(exmap & repl) const
+{
+	epvector s;
+	s.reserve(seq.size());
+	auto i = seq.begin(), end = seq.end();
+	while (i != end) {
+		s.push_back(split_ex_to_pair(recombine_pair_to_ex(*i).to_rational(repl)));
+		++i;
+	}
+	ex oc = overall_coeff.to_rational(repl);
+	if (oc.info(info_flags::numeric))
+		return thisexpairseq(s, overall_coeff);
+	else
+		s.push_back(combine_ex_with_coeff_to_pair(oc, _ex1));
+	return thisexpairseq(s, default_overall_coeff());
+}
+
+/** Implementation of ex::to_polynomial() for expairseqs. */
+ex expairseq::to_polynomial(exmap & repl) const
+{
+	epvector s;
+	s.reserve(seq.size());
+	auto i = seq.begin(), end = seq.end();
+	while (i != end) {
+		s.push_back(split_ex_to_pair(recombine_pair_to_ex(*i).to_polynomial(repl)));
+		++i;
+	}
+	ex oc = overall_coeff.to_polynomial(repl);
+	if (oc.info(info_flags::numeric))
+		return thisexpairseq(s, overall_coeff);
+	else
+		s.push_back(combine_ex_with_coeff_to_pair(oc, _ex1));
+	return thisexpairseq(s, default_overall_coeff());
+}
+
+
 /*
  *  Normal form of rational functions
  */
@@ -593,6 +786,9 @@ ex ex::numer_denom() const
 		return e.subs(repl, subs_options::no_pattern);
 }
 
+//////////////////////////////////////////////////////////////////////
+// The following ex member functions assume ex is a pölynomial.
+// They should be moved into mpoly-ginac.cpp and declared only ifndef GIAC
 
 /*
  *  Computation of LCM of denominators of coefficients of a polynomial
@@ -635,6 +831,269 @@ numeric lcm_of_coefficients_denominators(const ex &e)
 {
 	return lcmcoeff(e, *_num1_p);
 }
+
+/** Return maximum (absolute value) coefficient of a polynomial.
+ *  This function is used internally by heur_gcd().
+ *
+ *  @return maximum coefficient
+ *  @see heur_gcd */
+numeric ex::max_coefficient() const
+{
+	return bp->max_coefficient();
+}
+
+/** Implementation ex::max_coefficient().
+ *  @see heur_gcd */
+numeric basic::max_coefficient() const
+{
+	return *_num1_p;
+}
+
+numeric numeric::max_coefficient() const
+{
+	return abs();
+}
+
+numeric add::max_coefficient() const
+{
+	auto it = seq.begin();
+	auto itend = seq.end();
+	GINAC_ASSERT(is_exactly_a<numeric>(overall_coeff));
+	numeric cur_max = abs(ex_to<numeric>(overall_coeff));
+	while (it != itend) {
+		numeric a;
+		GINAC_ASSERT(!is_exactly_a<numeric>(it->rest));
+		a = abs(ex_to<numeric>(it->coeff));
+		if (a > cur_max)
+			cur_max = a;
+		it++;
+	}
+	return cur_max;
+}
+
+numeric mul::max_coefficient() const
+{
+#ifdef DO_GINAC_ASSERT
+	epvector::const_iterator it = seq.begin();
+	epvector::const_iterator itend = seq.end();
+	while (it != itend) {
+		GINAC_ASSERT(!is_exactly_a<numeric>(recombine_pair_to_ex(*it)));
+		it++;
+	}
+#endif // def DO_GINAC_ASSERT
+	GINAC_ASSERT(is_exactly_a<numeric>(overall_coeff));
+	return abs(ex_to<numeric>(overall_coeff));
+}
+
+
+/** Apply symmetric modular homomorphism to an expanded multivariate
+ *  polynomial.  This function is usually used internally by heur_gcd().
+ *
+ *  @param xi  modulus
+ *  @return mapped polynomial
+ *  @see heur_gcd */
+ex basic::smod(const numeric &xi) const
+{
+	return *this;
+}
+
+ex numeric::smod(const numeric &xi) const
+{
+	return GiNaC::smod(*this, xi);
+}
+
+ex add::smod(const numeric &xi) const
+{
+	epvector newseq;
+	newseq.reserve(seq.size()+1);
+	auto it = seq.begin();
+	auto itend = seq.end();
+	while (it != itend) {
+		GINAC_ASSERT(!is_exactly_a<numeric>(it->rest));
+		numeric num_coeff = GiNaC::smod(ex_to<numeric>(it->coeff), xi);
+		if (!num_coeff.is_zero())
+			newseq.push_back(expair(it->rest, num_coeff));
+		it++;
+	}
+	GINAC_ASSERT(is_exactly_a<numeric>(overall_coeff));
+	numeric num_coeff = GiNaC::smod(ex_to<numeric>(overall_coeff), xi);
+	return (new add(newseq, num_coeff))->setflag(status_flags::dynallocated);
+}
+
+ex mul::smod(const numeric &xi) const
+{
+#ifdef DO_GINAC_ASSERT
+	epvector::const_iterator it = seq.begin();
+	epvector::const_iterator itend = seq.end();
+	while (it != itend) {
+		GINAC_ASSERT(!is_exactly_a<numeric>(recombine_pair_to_ex(*it)));
+		it++;
+	}
+#endif // def DO_GINAC_ASSERT
+	auto  mulcopyp = new mul(*this);
+	GINAC_ASSERT(is_exactly_a<numeric>(overall_coeff));
+	mulcopyp->overall_coeff = GiNaC::smod(ex_to<numeric>(overall_coeff),xi);
+	mulcopyp->clearflag(status_flags::evaluated);
+	mulcopyp->clearflag(status_flags::hash_calculated);
+	return mulcopyp->setflag(status_flags::dynallocated);
+}
+
+/*
+ *  Separation of unit part, content part and primitive part of polynomials
+ */
+
+/** Compute unit part (= sign of leading coefficient) of a multivariate
+ *  polynomial in Q[x]. The product of unit part, content part, and primitive
+ *  part is the polynomial itself.
+ *
+ *  @param x  main variable
+ *  @return unit part
+ *  @see ex::content, ex::primpart, ex::unitcontprim */
+ex ex::unit(const ex &x) const
+{
+	ex c = expand().lcoeff(x);
+	if (is_exactly_a<numeric>(c))
+		return c.info(info_flags::negative) ?_ex_1 : _ex1;
+	else {
+		ex y;
+		if (get_first_symbol(c, y))
+			return c.unit(y);
+		else
+			throw(std::invalid_argument("invalid expression in unit()"));
+	}
+}
+
+
+/** Compute content part (= unit normal GCD of all coefficients) of a
+ *  multivariate polynomial in Q[x]. The product of unit part, content part,
+ *  and primitive part is the polynomial itself.
+ *
+ *  @param x  main variable
+ *  @return content part
+ *  @see ex::unit, ex::primpart, ex::unitcontprim */
+ex ex::content(const ex &x) const
+{
+	if (is_exactly_a<numeric>(*this))
+		return info(info_flags::negative) ? -*this : *this;
+
+	ex e = expand();
+	if (e.is_zero())
+		return _ex0;
+
+	// First, divide out the integer content (which we can calculate very efficiently).
+	// If the leading coefficient of the quotient is an integer, we are done.
+	ex c = e.integer_content();
+	ex r = e / c;
+	int deg = r.degree(x);
+	ex lcoef = r.coeff(x, deg);
+	if (lcoef.info(info_flags::integer))
+		return c;
+
+	// GCD of all coefficients
+	int ldeg = r.ldegree(x);
+	if (deg == ldeg)
+		return lcoef * c / lcoef.unit(x);
+	ex cont = _ex0;
+	for (int i=ldeg; i<=deg; i++)
+		cont = gcd(r.coeff(x, i), cont, nullptr, nullptr, false);
+	return cont * c;
+}
+
+
+/** Compute primitive part of a multivariate polynomial in Q[x]. The result
+ *  will be a unit-normal polynomial with a content part of 1. The product
+ *  of unit part, content part, and primitive part is the polynomial itself.
+ *
+ *  @param x  main variable
+ *  @return primitive part
+ *  @see ex::unit, ex::content, ex::unitcontprim */
+ex ex::primpart(const ex &x) const
+{
+	// We need to compute the unit and content anyway, so call unitcontprim()
+	ex u, c, p;
+	unitcontprim(x, u, c, p);
+	return p;
+}
+
+
+/** Compute primitive part of a multivariate polynomial in Q[x] when the
+ *  content part is already known. This function is faster in computing the
+ *  primitive part than the previous function.
+ *
+ *  @param x  main variable
+ *  @param c  previously computed content part
+ *  @return primitive part */
+ex ex::primpart(const ex &x, const ex &c) const
+{
+	if (is_zero() || c.is_zero())
+		return _ex0;
+	if (is_exactly_a<numeric>(*this))
+		return _ex1;
+
+	// Divide by unit and content to get primitive part
+	ex u = unit(x);
+	if (is_exactly_a<numeric>(c))
+		return *this / (c * u);
+	else
+		return quo(*this, c * u, x, false);
+}
+
+
+/** Compute unit part, content part, and primitive part of a multivariate
+ *  polynomial in Q[x]. The product of the three parts is the polynomial
+ *  itself.
+ *
+ *  @param x  main variable
+ *  @param u  unit part (returned)
+ *  @param c  content part (returned)
+ *  @param p  primitive part (returned)
+ *  @see ex::unit, ex::content, ex::primpart */
+void ex::unitcontprim(const ex &x, ex &u, ex &c, ex &p) const
+{
+	// Quick check for zero (avoid expanding)
+	if (is_zero()) {
+		u = _ex1;
+		c = p = _ex0;
+		return;
+	}
+
+	// Special case: input is a number
+	if (is_exactly_a<numeric>(*this)) {
+		if (info(info_flags::negative)) {
+			u = _ex_1;
+			c = abs(ex_to<numeric>(*this));
+		} else {
+			u = _ex1;
+			c = *this;
+		}
+		p = _ex1;
+		return;
+	}
+
+	// Expand input polynomial
+	ex e = expand();
+	if (e.is_zero()) {
+		u = _ex1;
+		c = p = _ex0;
+		return;
+	}
+
+	// Compute unit and content
+	u = unit(x);
+	c = content(x);
+
+	// Divide by unit and content to get primitive part
+	if (c.is_zero()) {
+		p = _ex0;
+		return;
+	}
+	if (is_exactly_a<numeric>(c))
+		p = *this / (c * u);
+	else
+		p = quo(e, c * u, x, false);
+}
+
+
 
 
 } // namespace GiNaC
